Move TopUniqueN to it's own module
